project: Apply more C++ paradigms to the code
- Use auto in places where code clarity is improved or identical.
- Replace trivial constructors and destructors with default.
- Use true random for random generation.
- Use std::string_view where it is valid to do so.
- Apply const where it is valid to do so.
- Use references where it is valid to do so.
- Manually optimize memory usage with std::move and std::copy.
- Opt for memory efficient containers where the size is known ahead of time.

+++ source/nvidia/ar/nvidia-ar-facedetection.cpp
@@ -92,7 +92,7 @@ void ar::facedetection::set_tracking_limit(size_t v)
 
 	// Update bounding boxes structure.
 	_bboxes.rects   = _rects.data();
-	_bboxes.maximum = v;
+	_bboxes.maximum = static_cast<uint8_t>(v);
 	_bboxes.current = 0;
 
 	// Update feature.
